ai-intelligence-layer/requirements.txt
# AI Intelligence Layer - Python Dependencies
# All models run locally, no API calls needed

# Core dependencies
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0

# HuggingFace Transformers
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0

# Optional: Accelerate for faster loading
accelerate==0.25.0

# Image processing
opencv-python==4.8.1.78

# Note: Models will be downloaded automatically on first use
# They are cached in ~/.cache/huggingface/
